requirements.txt
flask
strawberry-graphql
sqlalchemy
psycopg2
alembic